move SigNameVisitor out of always_comb
